+++ Changes.md
@@ -7,7 +7,7 @@ available [here](https://gitlab.com/CDAGaming/CraftPresence/-/compare/release%2F
 
 ### Changes
 
-* Added support for the `|` (OR) Operator in `Presence Settings` fields
+* Added support for the `|` (OR) Operator in `Presence Settings` fields, if `allowPlaceholderOperators` is true
     * This change can be used to combine 2 placeholders, in which the mod will choose the first
       non-null placeholder. One such example is that doing `&SERVER&|&PACK&` will show the pack message, if the server
       message
@@ -17,12 +17,14 @@ available [here](https://gitlab.com/CDAGaming/CraftPresence/-/compare/release%2F
       them automatically applied upon config regeneration.
     * This can also be considered a fix for when using a server icon while a pack icon was also available, which is a
       scenario that can happen under default behavior, and the reasoning for this addition.
+* Added `allowPlaceholderPreviews` as an Advanced Setting to replace needing `Debug Mode` for placeholder previews
 
 ### Fixes
 
 * Backend: Fixed shadowed dependencies breaking Fabric and Forge Developer Environment functionality
 * Backend: Fixed an issue where `getArgumentsMatching` could sometimes lead to unintended behavior
 * Backend: Fixed a NPE from `generateArgumentMessage` if it is fired before the IPC instance is initialized
+* Fixed an regression causing null characters to be considered valid for the `splitCharacter` setting
 * Fixed possible NPEs within `ServerUtils` if `currentServerMessage` or `currentServerIcon` was null
     * This also fixes compatibility with Replay Mod and similar mods